+++ lib/main.dart
@@ -6,16 +6,15 @@ import 'dart:core';
 import 'package:flutter/material.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
-import 'package:httapp/services/local_file_handling.dart';
-import 'package:httapp/services/text_theme_manager.dart';
 
 // pub.dev packages
 import 'package:yaml/yaml.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:flutter_native_splash/flutter_native_splash.dart';
+import 'package:latlong2/latlong.dart';
 
 // httapp
-import 'package:httapp/ui/main_navigation.dart';
+import 'package:httapp/pages/main_navigation.dart';
 import 'package:httapp/ui/theme.dart';
 // httapp models
 import 'package:httapp/models/geojson_parse.dart';
@@ -25,6 +24,8 @@ import 'package:httapp/models/local_path.dart';
 import 'package:httapp/services/firebase/remote_file_handling.dart';
 import 'package:httapp/services/firebase/remote_asset_methods.dart';
 import 'package:httapp/services/version_service.dart';
+import 'package:httapp/services/local_file_handling.dart';
+import 'package:httapp/services/text_theme_service.dart';
 
 // firebase api & options
 import 'package:firebase_storage/firebase_storage.dart';
@@ -39,13 +40,15 @@ import 'firebase_options.dart';
 /// Used for parsing from tree_manifest.yaml
 var treeIdMap = {};
 
-/// Global Map from [id] to the Map {'id': String, 'name': String, 'body': String, 'imageFileList': List`<`File`>`, 'thmFile': File}
+/// Global Map from [id] to the Map 
+/// {'id': String, 'name': String, 'body': String, 'latlong': LatLng,
+/// 'imageFileList': List`<`File`>`, 'thmFile': File}
 /// 
 /// This is for the [TreeTemplatePage] constructor
 Map<String, Map<String, dynamic >> treePageData = {};
 
 /// Global list of POIs
-List<PointOfInterest> poiData = [];
+Map<String, PointOfInterest> poiData = {};
 
 List<PointOfInterest> interpretiveSignData = [];
 
@@ -73,103 +76,16 @@ Future<void> main() async {
   /// Initialize text theme service to manage text across the app during runtime
   TextThemeService().init(GoogleFonts.latoTextTheme());
 
+  /// Initialize mapController service to handle the map controller
+  MapControllerService().init();
+
   // Initialize Firebase, utilizing the firebase_options.dart file that is generated by FlutterFire CLI.
   await Firebase.initializeApp(options: DefaultFirebaseOptions.currentPlatform);
 
   // create asset handler for downloading data from server
   RemoteAssets assetHandler = RemoteAssets(instance: FirebaseStorage.instance, downloadingFilesSet: downloadingFiles);
 
-  /*
-    Activate App Check with automatic provider selection.
-    If built with debug mode, FirebaseAppCheck will use the debug provider.
-    Otherwise it will use the platform specific provider.
-  */
-  try {
-    await FirebaseAppCheck.instance.activate(
-      // On Android, use debug provider in debug mode, Play Integrity in release
-      providerAndroid: kDebugMode
-          ? AndroidDebugProvider()
-          : AndroidPlayIntegrityProvider(),
-      
-      // On iOS, use debug provider in debug mode, App Attest in release
-      providerApple: kDebugMode
-        ? AppleDebugProvider()
-        : AppleAppAttestProvider(),
-    );
-  } catch(e) {
-    debugPrint('[main] Exception during Appcheck activate: $e');
-    rethrow;
-  }
-  /*
-    Download and/or read the tree_manifest.yaml file as a string.
-    Parse the string and return it as a YamlList.
-    Append the results of the YamlList to treeIdMap, Map<dynamic, dynamic>
-  */
-  try {
-    // load the tree manifest file from local or remote
-    yamlString = await assetHandler.loadManifest(treeManifestFile);
-
-    // read the YAML Stream
-    var treeDoc = loadYamlStream(yamlString);
-
-    // collapse the YamlList into a Map<dynamic, dynamic>
-    for (var entry in treeDoc) {
-      treeIdMap.addAll(entry);
-    }
-  }
-  catch(e) {
-    // TODO better error handling than this
-    print("An error occured while parsing the YAML manifest!");
-    print(e);
-  }
-  
-  /*
-    create treePageData entries for expected trees in the yaml manifest
-  */
-  try{
-    // save a ListResult to reference during loop
-    ListResult resultList = await RemoteFileHandler.listRemoteImageFiles(FirebaseStorage.instance);
-    imageListResult = resultList.items.map((ref) => File(ref.fullPath)).toList();
-    localImageList = false;
-    
-
-  } catch(e) {
-    debugPrint('istRemoteImageFiles exception');
-    if (imageListResult == null) {
-      imageListResult = LocalFileHandler.listFiles(await localImagePath);
-      localImageList = true;
-    }
-  }
-
-  try {
-    for (var id in treeIdMap.keys) {
-      // retrieve the description, images, and thumbnails
-      var descriptionString = await assetHandler.treeDescription(id);   // description is pulled from internet here
-
-      // Load the thumbnail either from a local file or downloaded from the server
-      // Stored in memory as a `Uint8List` so `Image.memory()` may be used instead of `Image.file()`.
-      // Results in more memory usage, but much better performance because the app doesn't wait for slow disk read.
-      Uint8List? thmFileBytes = await assetHandler.loadThumbnailFile(id);
-
-      List<File> imageFileList = await assetHandler.treeImgFileList(id, imageListResult!, isLocal: localImageList);   // image is pulled from internet here
-      
-
-      print('[main][treePageData $id] imageFileList: $imageFileList');
-      
-      // add the matching remote data to the 
-      treePageData[id] = {'id': id, 'name': treeIdMap[id], 'body': descriptionString, 'imageFileList': imageFileList, 'thumbnail': thmFileBytes};
-    }
-  }
-  catch(e) {
-    // TODO better error handling than this
-    print("An error occured while building tree info pages!");
-    print(e);
-  }
-
-  /*
-    create POI marker list for map
-  */
-
+  /* Temporary, example, data */
   String signsGeoJSON= 
   '''
 {
@@ -223,12 +139,113 @@ Future<void> main() async {
 }
 
   ''';
+
+  /*
+    Activate App Check with automatic provider selection.
+    If built with debug mode, FirebaseAppCheck will use the debug provider.
+    Otherwise it will use the platform specific provider.
+  */
+  try {
+    await FirebaseAppCheck.instance.activate(
+      // On Android, use debug provider in debug mode, Play Integrity in release
+      providerAndroid: kDebugMode
+          ? AndroidDebugProvider()
+          : AndroidPlayIntegrityProvider(),
+      
+      // On iOS, use debug provider in debug mode, App Attest in release
+      providerApple: kDebugMode
+        ? AppleDebugProvider()
+        : AppleAppAttestProvider(),
+    );
+  } catch(e) {
+    debugPrint('[main] Exception during Appcheck activate: $e');
+    rethrow;
+  }
+  /*
+    Download and/or read the tree_manifest.yaml file as a string.
+    Parse the string and return it as a YamlList.
+    Append the results of the YamlList to treeIdMap, Map<dynamic, dynamic>
+  */
+  try {
+    // load the tree manifest file from local or remote
+    yamlString = await assetHandler.loadManifest(treeManifestFile);
+
+    // read the YAML Stream
+    var treeDoc = loadYamlStream(yamlString);
+
+    // collapse the YamlList into a Map<dynamic, dynamic>
+    for (var entry in treeDoc) {
+      treeIdMap.addAll(entry);
+    }
+  }
+  catch(e) {
+    // TODO better error handling than this
+    print("An error occured while parsing the YAML manifest!");
+    print(e);
+  }
+  
+  // Create list of images
+  try{
+    // 
+    ListResult resultList = await RemoteFileHandler.listRemoteImageFiles(FirebaseStorage.instance);
+
+    //
+    imageListResult = resultList.items.map((ref) => File(ref.fullPath)).toList();
+    localImageList = false;
+  } catch(e) {
+    debugPrint('istRemoteImageFiles exception');
+
+    //
+    if (imageListResult == null) {
+
+      //
+      imageListResult = LocalFileHandler.listFiles(await localImagePath);
+      //
+      
+      localImageList = true;
+    }
+  }
+
+  // read the GeoJSON data for the various types of points
   try {
     poiData = GeojsonParse.parsePointOfInterests(debugGeoJSON);
-    interpretiveSignData = GeojsonParse.parsePointOfInterests(signsGeoJSON);
+    interpretiveSignData = GeojsonParse.parsePointOfInterests(signsGeoJSON).values.toList();
   } catch(e) {
     // TODO better error handling than this
     print("An error occured while parsing GeoJSON!");
+    print(e);
+  }
+
+
+  // load data for tree information pages
+  try {
+    for (var id in treeIdMap.keys) {
+      // retrieve the description, images, and thumbnails
+      var descriptionString = await assetHandler.treeDescription(id);   // description is pulled from internet here
+
+      // Load the thumbnail either from a local file or downloaded from the server
+      // Stored in memory as a `Uint8List` so Image.memory() may be used instead of Image.file()
+      Uint8List? thmFileBytes = await assetHandler.loadThumbnailFile(id);
+
+      List<File> imageFileList = await assetHandler.treeImgFileList(id, imageListResult!, isLocal: localImageList);   // image is pulled from internet here
+      
+
+      print('[main][treePageData $id] imageFileList: $imageFileList');
+      
+      // add the matching remote data to the 
+      treePageData[id] = {
+        'id': id,
+        'name': treeIdMap[id],
+        'body': descriptionString,
+        //'latlng': 
+        'imageFileList': imageFileList,
+        'thumbnail': thmFileBytes
+        };
+    }
+  }
+  catch(e) {
+    // TODO better error handling than this
+    print("An error occured while building tree info pages!");
     print(e);
   }
