+++ lib/pages/settings/customization_settings.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:httapp/main.dart';
-import 'package:httapp/services/text_theme_manager.dart';
+import 'package:httapp/services/text_theme_service.dart';
 
 class CustomizationSettings extends StatelessWidget {
   final String title = 'Customization Settings';
